Updates native bridge module to emit JS events
